% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGaslighting.R
\name{estimateGaslighting}
\alias{estimateGaslighting}
\title{estimateGaslighting}
\usage{
estimateGaslighting(textData)
}
\arguments{
\item{textData}{a text data frame meeting format requirements}
}
\value{
a table of cosine distances one for every document
}
\description{
Computes cosine distance from each input text to the base gaslighting vector
}
