% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaslight_summarize.R
\name{gaslight_summarize}
\alias{gaslight_summarize}
\title{gaslight_summarize}
\usage{
gaslight_summarize(x)
}
\arguments{
\item{gaslight_prepped}{a dataframe generated by the gaslight_prep() function}
}
\value{
a table of cosine distances one for every document
}
\description{
Computes cosine distance from each input text to the base gaslighting vector
}
