% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaslight_prep.R
\name{prep_txt}
\alias{prep_txt}
\title{clean_me}
\usage{
prep_txt(x)
}
\arguments{
\item{clean_dat}{a dataframe generated by the read_me() function}
}
\value{
a dataframe one-word-per-row format with variables of interest appended
}
\description{
Cleans and format input text, omit stopwords, lemmatize, splits to one word per row
}
